feat(core): use display-buffer to show results
Display buffer is better choice than pop-to-buffer because most of the
time the user wants to edit the query and see the result somewhere
else without being forced to switch the active window.

Also display-buffer is customizable via display-buffer-alist settings,
so this way users have more freedom on how exactly results are
displayed.